Add files via upload
I am using php and mysql, mysqli to be exact and making a login and signup system which allows users to create accounts as well as login to the account. Im using different techniques to avoid password leakages, sql injections and other critical security updates are here. Also form input validation is there. We are using mysql database to keep the user infomations. Once the user logged in, the navigation bar changes accordingly and adding user's name in to website body as well as nav bar